+++ example/lib/router/routes.dart
@@ -6,4 +6,5 @@ class Routes {
   static const cameraRecognitionPage = '/cameraRecognitionPage';
   static const foodSearchPage = '/foodSearchPage';
   static const staticImagePage = '/staticImagePage';
+  static const suggestionsPage = '/suggestionsPage';
 }