+++ lib/src/models/passio_id_entity_types.dart
@@ -1,9 +1,0 @@
-/// Defines different categories of food items provided by the SDK.
-enum PassioIDEntityType {
-  group,
-  item,
-  recipe,
-  barcode,
-  packagedFoodCode,
-  nutritionFacts
-}